pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "documind-rag"
version = "1.0.0"
description = "Production-ready RAG over PDFs: ask questions, get answers with cited sources."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "DocuMind-RAG" }]
keywords = ["rag", "retrieval", "pdf", "embeddings", "faiss", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "types-requests>=2.28.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pypdf>=4.0.0",
    "tiktoken>=0.5.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "huggingface_hub>=0.25.0",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
documind = "documind.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "documind*"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "documind", "tests", "main.py", "build_index.py", "app_gradio.py"]
exclude = ["tests/sample_data", ".venv", "venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["src.*", "documind.*"]
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
